+++ src/utils/mod.rs
@@ -1,3 +1,4 @@
+pub mod env;
 pub mod git;
 pub mod path;
 pub mod process;